third_party/WebKit/LayoutTests/bluetooth/requestDevice/blacklisted-service-in-filter.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('BlacklistTestAdapter')
    .then(() => assert_promise_rejects_with_message(
      requestDeviceWithKeyDown({
        filters: [{services: ['human_interface_device']}]}),
      new DOMException(
        'requestDevice() called with a filter containing a blacklisted UUID. ' +
        'https://goo.gl/4NeimX',
        'SecurityError'),
      'Requesting blacklisted service rejects.'));
}, 'Reject with SecurityError if requesting a blacklisted service.');
</script>
